requirements.txt
asgiref==3.3.1
Django==3.1.5
django-environ==0.4.5
django-redis==4.12.1
django-summernote==0.8.11.6
django-webpack-loader==0.7.0
gunicorn==20.0.4
install==1.3.4
MarkupSafe==1.1.1
psycopg2==2.8.6
pytz==2020.5
redis==3.5.3
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.3
